+++ app/src/main/java/me/portailler/florian/testanimation/ui/designs/state/DesignNavProvider.kt
@@ -1,5 +1,6 @@
 package me.portailler.florian.testanimation.ui.designs.state
 
+import me.portailler.florian.testanimation.ui.designs.customBrush.SquircleScreen
 import me.portailler.florian.testanimation.ui.designs.quarter.QuarterScreen
 
 object DesignNavProvider {
@@ -10,7 +11,12 @@ object DesignNavProvider {
 				title = "Quarter",
 				description = "A quarter of a circle",
 				screen = QuarterScreen()
-			)
+			),
+			DesignNavState(
+				title = "Squicle",
+				description = "An example of a squircle",
+				screen = SquircleScreen()
+			),
 		)
 	}
 }