+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- feat(Squircle): added squircle shape dependency
 - feat(App icon): change app icon dynamically
 - feat(Tab bar): build custom tab bar Composable
 - feat(shortcut): Added `ShortcutActivity` to test application shortcuts on smartphone home screen